+++ src/assets/mocks/userProfile.ts
@@ -1,0 +1,29 @@
+export const userProfile = {
+  name: '홍길동',
+  email: 'user@example.com',
+  nickname: '보드게임',
+  profile_image: '',
+  review_count: 14,
+  like_count: 25,
+  created_at: '2025-07-22T00:00:00Z',
+  preferred_genres: [
+    '전략',
+    '카드',
+    '추리',
+    '파티',
+    '블러핑',
+    '협동',
+    '경제',
+    '문명',
+    '정치',
+    '타일 배치',
+    '주사위',
+    '세트컬렉션',
+    '거래',
+    '배틀',
+    '영역확장',
+    '스토리텔링',
+  ],
+  preferred_playtimes: ['30분 미만', '1시간 이상'],
+  popular_genres: ['전략', '파티', '추리'],
+};
